refactor(broker): step 1/6 — move replay-check submodules into a directory
Step 1 of decomposing replay-check.ts into cost-ledger/replay-check/.
Moves the two reactive extractions into the new directory so the
target structure is rooted before splitting the orchestrator file:

  cost-ledger/replay-check/
  ├── budget-candidate-index.ts  (moved, git mv preserves history)
  └── unsafe-lifetime-accumulator.ts  (moved)

Internal imports updated: type-only imports of ReplayedBudget /
ReplayDiscrepancy point to ../replay-check.ts; the inline re-export
in replay-check.ts points to ./replay-check/budget-candidate-index.ts.

All 249 broker tests pass; typecheck clean.
